requirements.txt
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyYAML==6.0.1
requests==2.32.3
urllib3==2.2.2
